requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.1.0
transformers>=4.49.0
albumentations>=1.3.0
pycocotools>=2.0.7
pyyaml>=6.0
numpy>=1.24.0
pillow>=10.0.0
scipy>=1.11.0
tqdm>=4.65.0

# Monitoring
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
psutil>=5.9.0
pynvml>=11.5.0
aiosqlite>=0.20.0
websockets>=12.0
pydantic>=2.5.0
